Add files via upload
Finished all the menupoint icons (selfmade via inkscape) as well as make code more efficient. Also I started with the implementation of the actual pages that will be stacked via stackView. Also updated the apk so it matches the current code status. App can be installed on an android phone and be tested.