Add files via upload
The loader variant of the app has been replaced by soley using stackview for navigating the drawermenu. The "Diät festlegen" menu point has been started to develop. Now via a button click a predefined dietplan is being created dynamically as well as being able to destroy it to free space. In the next step I will implement a way to created multiple user defined dietplans as well as storing the current app state so that the data isn't being lost after app is closed. The apk has been updated accordingly and can be tested.